Move clip time indexer tests to proper test file
